Moved existing JSON output to 'pretty' to make way for using python json assembly routines (non-pretty).
